+++ changelog.md
@@ -13,6 +13,7 @@
 * Replaced the table `order` index with an `order, id` index to improve performance.
 * **breaking change** The API endpoint to rotate a form views slug has been moved to
   `/database/views/${viewId}/rotate-slug/`.
+* Increased maximum length of application name to 160 characters.
 
 ## Released (2021-11-25)